+++ src/torchmetrics/classification/exact_match.py
@@ -32,6 +32,7 @@ from torchmetrics.functional.classification.stat_scores import (
 )
 from torchmetrics.metric import Metric
 from torchmetrics.utilities.data import dim_zero_cat
+from torchmetrics.utilities.enums import ClassificationTaskNoBinary
 
 
 class MulticlassExactMatch(Metric):
@@ -117,6 +118,7 @@ class MulticlassExactMatch(Metric):
         )
 
     def update(self, preds, target) -> None:
+        """Update metric states with predictions and targets."""
         if self.validate_args:
             _multiclass_stat_scores_tensor_validation(
                 preds, target, self.num_classes, self.multidim_average, self.ignore_index
@@ -131,6 +133,7 @@ class MulticlassExactMatch(Metric):
             self.total += total
 
     def compute(self) -> Tensor:
+        """Compute metric."""
         correct = dim_zero_cat(self.correct) if isinstance(self.correct, list) else self.correct
         return _exact_match_reduce(correct, self.total)
 
@@ -249,6 +252,7 @@ class MultilabelExactMatch(Metric):
             self.total += total
 
     def compute(self) -> Tensor:
+        """Compute metric."""
         correct = dim_zero_cat(self.correct) if isinstance(self.correct, list) else self.correct
         return _exact_match_reduce(correct, self.total)
 
@@ -288,13 +292,14 @@ class ExactMatch:
         validate_args: bool = True,
         **kwargs: Any,
     ) -> Metric:
+        """Initialize task metric."""
+        task = ClassificationTaskNoBinary.from_str(task)
         kwargs.update(
             {"multidim_average": multidim_average, "ignore_index": ignore_index, "validate_args": validate_args}
         )
-        if task == "multiclass":
+        if task == ClassificationTaskNoBinary.MULTICLASS:
             assert isinstance(num_classes, int)
             return MulticlassExactMatch(num_classes, **kwargs)
-        if task == "multilabel":
+        if task == ClassificationTaskNoBinary.MULTILABEL:
             assert isinstance(num_labels, int)
             return MultilabelExactMatch(num_labels, threshold, **kwargs)
-        raise ValueError(f"Expected argument `task` to either be `'multiclass'` or `'multilabel'` but got {task}")