+++ src/torchmetrics/functional/multimodal/clip_score.py
@@ -17,11 +17,20 @@ import torch
 from torch import Tensor
 from typing_extensions import Literal
 
+from torchmetrics.utilities.checks import _in_doctest, _try_proceed_with_timeout
 from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
 
 if _TRANSFORMERS_AVAILABLE:
     from transformers import CLIPModel as _CLIPModel
     from transformers import CLIPProcessor as _CLIPProcessor
+
+    def _download_clip() -> None:
+        _CLIPModel.from_pretrained("openai/clip-vit-large-patch14")
+        _CLIPProcessor.from_pretrained("openai/clip-vit-large-patch14")
+
+    if _in_doctest() and not _try_proceed_with_timeout(_download_clip):
+        __doctest_skip__ = ["clip_score"]
+
 else:
     __doctest_skip__ = ["clip_score"]
     _CLIPModel = None  # type:ignore
@@ -97,9 +106,11 @@ def clip_score(
         "openai/clip-vit-large-patch14",
     ] = "openai/clip-vit-large-patch14",
 ) -> Tensor:
-    r"""`CLIP Score`_ is a reference free metric that can be used to evaluate the correlation between a generated
-    caption for an image and the actual content of the image. It has been found to be highly correlated with human
-    judgement. The metric is defined as:
+    r"""Calculates `CLIP Score`_ which is a text-to-image similarity metric.
+
+    CLIP is a reference free metric that can be used to evaluate the correlation between a generated caption for an
+    image and the actual content of the image. It has been found to be highly correlated with human judgement. The
+    metric is defined as:
 
     .. math::
         \text{CLIPScore(I, C)} = max(100 * cos(E_I, E_C), 0)